third_party/blink/web_tests/external/wpt/css/css-anchor-position/position-area-wm-dir-expected.txt

This is a testharness.js-based test.
[FAIL] Offsets for: self-x-start self-y-start with writing-mode / direction: containing-block: horizontal-tb / rtl
  assert_equals: Checking offsetLeft expected 0 but got 250
[FAIL] Offsets for: self-x-start self-y-start with writing-mode / direction: containing-block: vertical-lr / rtl
  assert_equals: Checking offsetTop expected 0 but got 225
[FAIL] Offsets for: self-x-start self-y-start with writing-mode / direction: containing-block: vertical-rl / ltr
  assert_equals: Checking offsetLeft expected 0 but got 250
[FAIL] Offsets for: self-x-start self-y-start with writing-mode / direction: containing-block: vertical-rl / rtl
  assert_equals: Checking offsetLeft expected 0 but got 250
[FAIL] Offsets for: self-x-start self-y-start with writing-mode / direction: self: horizontal-tb / rtl
  assert_equals: Checking offsetLeft expected 250 but got 0
[FAIL] Offsets for: self-x-start self-y-start with writing-mode / direction: self: vertical-lr / rtl
  assert_equals: Checking offsetTop expected 225 but got 0
[FAIL] Offsets for: self-x-start self-y-start with writing-mode / direction: self: vertical-rl / ltr
  assert_equals: Checking offsetLeft expected 250 but got 0
[FAIL] Offsets for: self-x-start self-y-start with writing-mode / direction: self: vertical-rl / rtl
  assert_equals: Checking offsetLeft expected 250 but got 0
Harness: the test ran to completion.
